fix: 2차 코드리뷰 반영
