+++ app/src/main/java/com/stormers/storm/base/BaseBindingFragment.kt
@@ -19,7 +19,7 @@ open class BaseBindingFragment<B : ViewDataBinding>(@LayoutRes private val layou
         savedInstanceState: Bundle?
     ): View {
         binding = DataBindingUtil.inflate(inflater, layoutRes, container, false)
-        binding.lifecycleOwner = this
+        binding.lifecycleOwner = viewLifecycleOwner
         return binding.root
     }
