+++ app/src/main/java/com/stormers/storm/mypage/MyPageService.kt
@@ -1,6 +1,6 @@
 package com.stormers.storm.mypage
 
-import com.stormers.storm.mypage.model.MyPageResponse
+import com.stormers.storm.mypage.model.ResponseMyPage
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Path
@@ -8,5 +8,5 @@ import retrofit2.http.Path
 interface MyPageService {
     @GET("/user/mypage/{user_idx}")
     fun getMypageData(
-        @Path("user_idx") userIdx : Int) : Call<MyPageResponse>
+        @Path("user_idx") userIdx : Int) : Call<ResponseMyPage>
 }
