+++ app/src/main/java/com/stormers/storm/mypage/datasource/MyPageDataSource.kt
@@ -1,11 +1,11 @@
 package com.stormers.storm.mypage.datasource
 
-import com.stormers.storm.mypage.model.MyPageResponse
+import com.stormers.storm.mypage.model.ResponseMyPage
 import okhttp3.ResponseBody
 
 interface MyPageDataSource {
     interface MyPageCallback{
-        fun onSuccessFetchUserData(myPageResponse: MyPageResponse)
+        fun onSuccessFetchUserData(responseMyPage: ResponseMyPage)
         fun onFailedFetchUserData(networkError: Boolean, errorBody: ResponseBody?)
     }
