+++ app/src/main/java/com/stormers/storm/mypage/fragment/MyPageFragment.kt
@@ -1,6 +1,7 @@
 package com.stormers.storm.mypage.fragment
 
 import android.os.Bundle
+import android.util.Log
 import android.view.View
 import android.widget.Toast
 import androidx.fragment.app.viewModels
@@ -17,7 +18,6 @@ class MyPageFragment : BaseBindingFragment<FragmentMyPageBinding>(R.layout.fragm
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         binding.viewModel = viewModel
-        binding.lifecycleOwner = viewLifecycleOwner
 
         setToolbar()
     }
