+++ app/src/main/java/com/stormers/storm/mypage/model/ResponseMyPage.kt
@@ -1,6 +1,6 @@
 package com.stormers.storm.mypage.model
 
-data class MyPageResponse(
+data class ResponseMyPage(
     val status : Int,
     val success : Boolean,
     val message : String,
