+++ app/src/main/res/layout/fragment_my_page.xml
@@ -93,6 +93,7 @@
             tools:text="이영진"/>
 
         <androidx.appcompat.widget.AppCompatButton
+            android:id="@+id/button_rename"
             android:layout_width="0dp"
             android:layout_height="0dp"
             app:layout_constraintDimensionRatio="1:1"
